mcps/summarization-mcp/pyproject.toml
[tool.coverage.run]
branch = true
source = ["summarization-mcp"]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = """
 --cov
 --cov-report html:'../coverage/summarization-mcp/html'
 --cov-report xml:'../coverage/summarization-mcp/coverage.xml'
 --html='../reports/summarization-mcp/unittests/html/index.html'
 --junitxml='../reports/summarization-mcp/unittests/junit.xml'
"""

[project]
name = "summarization-mcp"
version = "1.0.0"
description = "Automatically generated by Nx."
requires-python = ">=3.12,<4"
readme = 'README.md'
dependencies = [
    "python-dotenv>=1.0.0",
    "fastmcp>=2.11.3",
    "motor>=3.7.1",
    "regex>=2024.11.6",
    "semchunk>=3.2.2",
    "tiktoken>=0.4.0",
    "lingua-language-detector>=2.1.1",
    "litellm>=1.74.9.post1",
]

[tool.hatch.build.targets.wheel]
packages = ["summarization-mcp"]

[tool.hatch.metadata]
allow-direct-references = true

[dependency-groups]
dev = [
    "autopep8>=2.3.1",
    "ruff>=0.8.2",
    "pytest>=8.3.4",
    "pytest-sugar>=1.0.0",
    "pytest-cov>=6.0.0",
    "pytest-html>=4.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
exclude = [".ruff_cache", ".svn", ".tox", ".venv", "dist"]

line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
ignore = []

fixable = ["ALL"]
unfixable = []
